+++ R/vctrs-math.R
@@ -18,7 +18,7 @@ vec_math_bigfloat <- function(.fn, .x, ..., na.rm = FALSE) {
     cummin = c_bigfloat_cummin(.x),
     exp = c_bigfloat_exp(.x),
     expm1 = c_bigfloat_expm1(.x),
-    log = c_bigfloat_log(.x),
+    log = bigfloat_log(.x, ...),
     log10 = c_bigfloat_log10(.x),
     log2 = c_bigfloat_log2(.x),
     log1p = c_bigfloat_log1p(.x),
@@ -86,4 +86,14 @@ vec_math.bignum_biginteger <- function(.fn, .x, ..., na.rm = FALSE) {
     # else
     vec_math_bigfloat(.fn, .x, ..., na.rm = na.rm)
   )
+}
+
+# functions with unnamed arguments require special handling --------------------
+
+bigfloat_log <- function(x, base) {
+  if (is_missing(base)) {
+    c_bigfloat_log(x)
+  } else {
+    c_bigfloat_log(x) / c_bigfloat_log(vec_cast(base, bigfloat()))
+  }
 }